extract api call to static helper
