+++ app/src/main/java/com/nsc/covidscore/MainActivity.java
@@ -4,18 +4,18 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
 import android.util.Log;
+import android.widget.TextView;
 
-import com.android.volley.Request;
 import com.android.volley.RequestQueue;
-import com.android.volley.Response;
-import com.android.volley.VolleyError;
-import com.android.volley.toolbox.StringRequest;
 import com.nsc.covidscore.api.RequestSingleton;
+import com.nsc.covidscore.api.Requests;
+import com.nsc.covidscore.api.VolleyCallback;
 
 public class MainActivity extends AppCompatActivity {
     private static final String TAG = MainActivity.class.getSimpleName();
     RequestQueue queue;
     RequestSingleton requestManager;
+    TextView textView;
 
 
     @Override
@@ -24,31 +24,16 @@ public class MainActivity extends AppCompatActivity {
         setContentView(R.layout.activity_main);
         requestManager = RequestSingleton.getInstance(this.getApplicationContext());
         queue = requestManager.getRequestQueue();
-        getCounty("King");
+        textView = findViewById(R.id.hello_world);
+        Requests.getCounty(this, "King", TAG, new VolleyCallback() {
+            @Override
+            public void getResponse(String response) {
+                Log.i(TAG, "getResponse: " + response);
+            }
+        });
         Log.d(TAG,"onCreate invoked");
     }
 
-    private void getCounty(String county) {
-        String url = "https://corona.lmao.ninja/v2/jhucsse/counties/" + county;
-
-        // Request a string response from the provided URL.
-        StringRequest stringRequest = new StringRequest(Request.Method.GET, url,
-                new Response.Listener<String>() {
-                    @Override
-                    public void onResponse(String response) {
-                        Log.i(TAG, "onResponse: " + response);
-                    }
-                }, new Response.ErrorListener() {
-            @Override
-            public void onErrorResponse(VolleyError error) {
-                Log.i(TAG, "onError: That didn't work!" + error);
-            }
-        });
-        stringRequest.setTag(TAG);
-
-        // Add the request to the RequestQueue.
-        requestManager.addToRequestQueue(stringRequest);
-    }
     @Override
     protected void onStop () {
         super.onStop();
@@ -56,5 +41,4 @@ public class MainActivity extends AppCompatActivity {
             requestManager.getRequestQueue().cancelAll(TAG);
         }
     }
-
 }
