+++ app/src/main/java/com/nsc/covidscore/api/Requests.java
@@ -1,0 +1,33 @@
+package com.nsc.covidscore.api;
+
+import android.content.Context;
+import android.util.Log;
+
+import com.android.volley.Request;
+import com.android.volley.Response;
+import com.android.volley.VolleyError;
+import com.android.volley.toolbox.StringRequest;
+
+public class Requests {
+    public static void getCounty(Context context, String county, String TAG, final VolleyCallback cb) {
+        String url = "https://corona.lmao.ninja/v2/jhucsse/counties/" + county;
+
+        // Request a string response from the provided URL.
+        StringRequest stringRequest = new StringRequest(Request.Method.GET, url,
+                new Response.Listener<String>() {
+                    @Override
+                    public void onResponse(String response) {
+                        cb.getResponse(response);
+                    }
+                }, new Response.ErrorListener() {
+            @Override
+            public void onErrorResponse(VolleyError error) {
+            }
+        });
+        stringRequest.setTag(TAG);
+
+        // Add the request to the RequestQueue.
+        RequestSingleton.getInstance(context.getApplicationContext()).addToRequestQueue(stringRequest);
+//        requestManager.addToRequestQueue(stringRequest);
+    }
+}
