+++ app/src/main/java/com/nsc/covidscore/api/VolleyCallback.java
@@ -1,0 +1,5 @@
+package com.nsc.covidscore.api;
+
+public interface VolleyCallback {
+    void getResponse(String response);
+}
